+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # CHANGELOG
 
+## 0.12.5
+
+`2019年3月4日`
+
+#### 新特性
+
+- 新增`ui-time`组件
+- `ui-grid`支持设置栅格列的位移，支持等宽栅格，支持水平和垂直对齐方式，支持设置栅格空隙尺寸
+
+#### 改进
+
+- 统一`ui-message`和`ui-load`组件中含义类似的配置名(`time`及`close-time`现在统一为`done-time`)
+
+<ui-label color="primary">升级注意事项</ui-label>
+
+- `ui-message`组件的`close-time`配置改名为`done-time`
+- `ui-load`组件的`time` 配置改名为`done-time`
+- 现在`ui-grid`中的`.c-[n]`改名为`.col-[n]`，且必须在`.row`中使用
+- `ui-grid`组件的`avggap`样式类改名为`space-around`
+- `ui-grid`组件的`nogap`样式类改名为`gutter-none`
+
 ## 0.12.4
 
 `2019年3月1日`